\name{tau-statistic-speedup-package}
\alias{tau-statistic-speedup-package}
\alias{tau-statistic-speedup}
\docType{package}
\title{\packageTitle{tau-statistic-speedup}}
\description{\packageDescription{tau-statistic-speedup}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{tau-statistic-speedup}
  \packageIndices{tau-statistic-speedup}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{tau-statistic-speedup}
  
  Maintainer: \packageMaintainer{tau-statistic-speedup}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
